feat: handle auto update for macos
