+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hydralauncher",
-  "version": "1.2.4",
+  "version": "1.2.3",
   "description": "Hydra",
   "main": "./out/main/index.js",
   "author": "Los Broxas",
