+++ src/locales/en/translation.json
@@ -29,7 +29,8 @@
     "downloads": "Downloads",
     "search_results": "Search results",
     "settings": "Settings",
-    "version_available": "Version {{version}} available. Click here to restart and install."
+    "version_available_install": "Version {{version}} available. Click here to restart and install.",
+    "version_available_download": "Version {{version}} available. Click here to download."
   },
   "bottom_panel": {
     "no_downloads_in_progress": "No downloads in progress",
