+++ src/main/events/autoupdater/check-for-updates.ts
@@ -15,13 +15,15 @@ const mockValuesForDebug = () => {
 };
 
 const checkForUpdates = async (_event: Electron.IpcMainInvokeEvent) => {
-  autoUpdater
-    .once("update-available", (info: UpdateInfo) => {
-      sendEvent({ type: "update-available", info });
-    })
-    .once("update-downloaded", () => {
+  autoUpdater.once("update-available", (info: UpdateInfo) => {
+    sendEvent({ type: "update-available", info });
+  });
+
+  if (process.platform !== "darwin") {
+    autoUpdater.once("update-downloaded", () => {
       sendEvent({ type: "update-downloaded" });
     });
+  }
 
   if (app.isPackaged) {
     autoUpdater.checkForUpdates();
