+++ src/main/events/autoupdater/restart-and-install-update.ts
@@ -1,13 +1,18 @@
 import { app } from "electron";
 import { registerEvent } from "../register-event";
 import updater from "electron-updater";
+import { releasesPageUrl } from "@main/constants";
 
 const { autoUpdater } = updater;
 
 const restartAndInstallUpdate = async (_event: Electron.IpcMainInvokeEvent) => {
   autoUpdater.removeAllListeners();
   if (app.isPackaged) {
-    autoUpdater.quitAndInstall(true, true);
+    if (process.platform === "darwin") {
+      open(`${releasesPageUrl}`);
+    } else {
+      autoUpdater.quitAndInstall(true, true);
+    }
   }
 };
